+++ frontend/package.json
@@ -7,6 +7,7 @@
     "bootstrap": "^4.3.1",
     "connected-react-router": "^6.5.2",
     "core-util-is": "^1.0.2",
+    "history": "^4.10.1",
     "react": "^16.10.1",
     "react-bootstrap": "^1.0.0-beta.14",
     "react-dom": "^16.10.1",
