+++ frontend/src/components/Collection/CollectionCard/CollectionCard.test.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-undef */
 import React from "react";
 import { shallow, mount } from "enzyme";
 import CollectionCard from "./CollectionCard";
