+++ frontend/src/components/Header/Header.test.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-undef */
 import React from "react";
 import { shallow } from "enzyme";
 import Header from "./Header";
