+++ frontend/src/store/store.js
@@ -6,7 +6,7 @@ import { createBrowserHistory } from "history";
 import rootReducer from "./reducers/index";
 
 export const history = createBrowserHistory();
-export const middlewares = [thunk, routerMiddleware(history)]
+export const middlewares = [thunk, routerMiddleware(history)];
 
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 const store = createStore(rootReducer(history), composeEnhancers(
